+++ addons/yat/src/resources/preferences/YatPreferences.cs
@@ -33,8 +33,9 @@ public partial class YatPreferences : Resource
 
 	#region Colors
 	[ExportGroup("Colors")]
-	[Export] public Color BackgroundColor { get; set; } = new("1d2229"); // #192229
 	[Export] public Color InputColor { get; set; } = new("15191fc0"); // #15191fc0
+	[Ignore]
+	[Export] public Color BackgroundColor { get; set; } = new("1d2229"); // #192229
 	[Export] public Color OutputColor { get; set; } = new("dfdfdf80"); // #dfdfdf80
 	[Export] public Color ErrorColor { get; set; } = new("ff786b"); // #ff7866
 	[Export] public Color WarningColor { get; set; } = new("ffdd65"); // #ffdd65