+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 -   '-e' option for 'cat' command.
 -   Ok to ECommandResult enum.
+-   Unavailable to ECommandResult enum.
 -   Ok method to ICommand interface.
 
 ### Changed