+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file.
 -   'cat' command by default uses FullWindowDisplay to display content of the file.
 -   'inspect' command displays node's UID.
 -   'InputColor' preference is used by YAT from now on.
+-   'BackgroundColor' preference is hidden.
 
 ### Fixed